admin/requirements.txt
pymongo
fastapi
uvicorn
opencage
jinja2
python-dotenv